Run more commands and with bigger content length
I'm just now realizing that there is a problem
with the way the test is structured. QuickCheck
does not always run the number of commands needed
to read all of the chunks from the FSM. This means
that the FSM stays around and are not killed
until they timeout. The test runs quickly enough that
we get a too many processes error if we run a lot of test
cases. More importantly, the test does not fully test
the FSM.