Start filling in the state callbacks
S24

I'm currently wondering if
there doesn't need to be a
distinction between add_new_manifest
and update_manifest. The rationale
is that both of them need to retrieve
whatever value is currently in Riak
and then can just "resolve" this
in-memory state with the value found.
If there is no value found, then
we simply create a new riak_object