Fix bucket creation when a content-type header with parameters is used.
Fixes #108

The webmachine resource for handling bucket creation requests was not
separating the media type from the parameters that were included
in the content-type header. This was causing the media type to be
unrecognized and resulting in 415 Unsupported Media Type errors
being returned.

Use mochiweb_util:parse_header to parse the content-type headers for
bucket and key resources.