Add EQC test to show that manifest resolution is commutative
S24

Add a test that (tries to) verifiy
that manifest sibling resolution
returns the same result, regardless
of the order of the siblings. It
was helpful in finding two "basic"
function clause bugs. More work
needs to be done to get manifests
that vary in more than just their
UUID and state.