Add module for marking manifests as deleted
riak_moss_delete_marker:delete/2 is
used to "mark" a manifest as deleted.
This is done before the actual blocks
are deleted. Once all of the blocks have
been deleted, the manifest transitions
into the deleted state (where it sticks
around for a bit longer, so that all
replicas have a chance to observe it).
This function currently doesn't spawn
any blocks GC, so that's a TODO.

S24