Parameterize BlockSize in every function
BlockSize was previously a variable
that was sometimes an argument, and sometimes
a global variable (because it was read from app.config).
This commit removes all of the places that it is read
from the app.config, _except_ for the creation
of put_fsm's in r_m_wm_key.erl. This makes the code
more pure, more resilient to block size changes,
and easier to test.