S175: Shorten KV bucket & key names for large file storage (LFS) blocks
Fixes #175 (partially).

Add a bucket_id to a new #moss_bucket_v2 record: this ID will be used
for the Riak KV bucket name for LFS blocks.  The result will be a fixed
size KV bucket + key name for LFS blocks.

There's a bug/mis-feature of the PB interface that doesn't appear to
transport the riak object's metadata dictionary across the "wire".  So my
efforts to add the S3 bucket & file name to the LFS block metadata
dictionary (as a "backreference" to give some human-friendly(ier) clue
about the S3 object the block belongs to) is effectively broken.