Usage Recording (fixes #6)
This is a squashed commit of the s6-access-log branch.

This commit encompasses the entirety of the "usage logging" feature.
HTTP access stats and storage stats are recording on a windowed-rollup
per-user basis.  HTTP access stats are gathered via a custom Webmachine
logger module.  Storage stats are gathered via MapReduce queries on the
manifest objects in each CS bucket.

Querying these stats is provided via a /usage resource via either query
param:

   GET /usage/USER_KEY_ID?a&b&s=START_TIME&e=END_TIME

or s3 object request:

   s3cmd get s3://usage/USER_KEY_ID/abj/START_TIME/END_TIME

Including the 'a' and 'b' option in each URL enables the inclusion of
access and storage (think "bytes") data, respectively, in the response.

The response is available in either JSON or XML format.  When using the
query-param URL, include an Accept header to choose JSON
(application/json) or XML (application/xml).  When using the s3 object
path, include 'j' for JSON or 'x' for XML in the component where 'a' and
'b' are specified.