Set content-type in context when using the default type.
Fixes GH143

If no content-type header is specified in a request the default
header type is not set in the key_context in the riak_moss_wm_key
webmachine resource. This causes a badarg error when the accept_body
function calls list_to_binary on the content-type from the context
and the value is undefined.