Remove automatic partial upload GC
It's been decided that reaping failed/partial
uploads is too dangerous to do automatically,
without substantial work. The basic reason is
that because of an extended netsplit, you might
end up GC the blocks for a write that you thought
had failed, but had actually completed on the other
side of the netsplit. This is not good if this is the
current active version of the file. We considered only
deleting them if we had observed an overwrite or delete
had happened after, but this would require using
vector clocks to determine "after", as we
can't trust wall clocks.

The first thought is that we'll provide
a tool that you can run to reap partially
written blocks, only to be run when there
are no netsplits. This might also be
dangerous though, as deployers are likely
to just run it in cron.