Use {notfound_ok, false} for r=1 block requests
Fixes #170

Without this patch, requests for
sufficiently large files will
fail when a single Riak node is
down. This happens (presumably) because
the fallback node is the quickest to
respond, and thus the block request
gets a 404. This commit adds
{notfound_ok, false} and {basic_quorum, false}
to the request so that the request
will only return 404 is N (3) vnodes
return 404.