don't bomb if the object being deleted doesn't exist
This case was making s3api_verification_test barf when the tests ran in a
different order than expected (because one of the tests attempted to
clean up something that shouldn't have been there).  Beyond the test
case, though, I think we don't want RCS to log an error when a user
attempts to delete something that doesn't exist.