Catch errors thrown from call to poolboy:checkout.
Fixes MOSS203

Add a catch in riak_moss_utils:riak_connection/1 to handle cases where
the call to poolboy:checkout/2 times out or otherwise throws an
error. The error is handled gracefully and an error code is returned
to the client that indicates the request should be retried, but that
the request rate is too high. Also add 10 second timeout for riak
connection establishment.