Add ping resource for Riak CS.
Fixes MOSS204

Add a webmachine resource for health checks of Riak CS. The resource
determines success of the ping request by attempting to acquire a Riak
protocol buffers connection and then calling the riakc_pb_socket:ping/2
function via that connection. If the Riak ping is successful a 200
is returned with 'OK' as the response body; otherwise, a 503 error
response code is returned.