Set metadata on PUT to avoid siblings exception
When a Riak object had siblings and you just update
the value, and don't supply metadata, a 'siblings' exception
will be raised on PUT. This patch adds a helper function
and use-of-it whenever we do a PUT where the previous
read could have had siblings. Since we don't (yet)
use Riak metadata, we just set the metadata to an empty
dict.