Add ability to configure the epoch start time for the GC daemon.
Fixes MOSS221

This is to address cases where a very large number of objects have
accumulated that need to be garbage collected such that it could
potentially overwhelm the system.  The epoch start can be adjusted so
that the backlog of objects can be handled in smaller batches and
hopefully avoid any system degradation.