Refactor handling of replies from delete FSM.
So, this refactoring tries to answer a couple of questions that
are yet unanswered ... so this patch may be introducing bugs.

%%   1. delete_fsm_pid=undefined is desirable in both ok & error cases?
%%   2. It's correct to *not* change pause_state?

This patch doesn't change pause_state in either riak_cs_gc_d FSM state,
because I don't understand why pause_state would be overwritten when
a reply is received while in 'paused' state but not in
'waiting_file_delete' state.

The original code is also doing some extra pattern matching when
in the paused state: pause_state=waiting_file_delete.  It looks like that's
an intentionally restrictive pattern match: there is no other clause for
handling delete FSM replies when pause_state /= waiting_file_delete.