Naming and type fixes for riak_cs_manifest_utils.erl
The name of riak_moss_manifest.erl bugged me, because it wasn't a
module that operated primarily on a single manifest record: it's
mostly operating on collections of manifests (all related to a
single S3 object).  So something like `riak_moss_manifest:new()`
should be creating a new manifest, merely looking at its name.
But that's not what `new()` does, so I've changed its name to
come closer to reflecting what it does: `new_dict()`.

Then I went on a bit of a orddict opacity rampage.  And I think I
found a type bug in `pending_delete_manifests()`: it was returning
{uuid(), ?MANIFEST} 2-tuples, and eventually
riak_cs_gc:move_manifests_to_gc_bucket ... AND IT ISN'T A TYPE ERROR
HERE AFTER ALL, FIX IT IN A NEW & SQUASHED
COMMIT......................................................................