Fix bone-headed block offset assumption, add more tests, all pass now!
I hard-coded a really stupid & wrong assumption about the block number
into a lot of code.  This commit fixes that assumption, adds more
tests (that check for the bogus assumption, t2()), and fix all of the
tests & other code broken by the change to 16 -> 32 bit integers for
the block number.