Fix issue with sibling creation on user record updates.
Fixes CS274

Vector clocks were not being used for user account updates and this
resulted in sibling creation and unexpected behavior from the user
perspective. This change reworks how user record updates are handled
so that now instead of storing only the vector clock in the webmachine
resource context and having to reconstruct an entire new object for an
update, we store the entire object in the context and make the
appropriate updates to that object prior to saving.