Don't crash in terminate if State isn't fully populated
Check to make sure that the pids we stop
in terminate/3 are actually pids before stopping them.
This was masking errors earlier in the FSM because crashes
in the terminate function were causing the previous error
to not be logged. Tested by adding some bugs calls to
foo:foo() in the various `prepare` function clauses.