finish refactoring authentication and bucket resources
* auth is done in the following 4 stages: identify -> lookup -> authenticate -> authorize. the identify stage is auth module specific, see the added identify/2 function in the riak_cs_auth
behaviour. identify/2 is expected to return the key for the user necessary to perform lookup as well as any arbitrary data that it does not wish to have to parse again, this data can be used
during authentication (e.g. returning signature from S3 Authorization header). The lookup is generic and is performed on the user key returned from identify using the riak client in the resource
context. authentication is auth module specific, like identify, and operates on the looked up user if there was one. if authentication succeeds we proceed to authorization. authorization is
submodule (wm resource) specific and a default is provided if the submodule does not have an authorize method.
* bucket resources (location, acl, objects, buckets, versioning)