Clean up and fix edge case in GC moving
Previously the manifests marked as scheduled
delete and moved to GC bucket would be any
in the `pending_delete` state. This was for
two reasons. One, `handle_mark_as_pending_delete`
would grab the `pending_delete` manifests based
on observing all the manifests, not just the
specific UUIDs passed in. Two, `pending_delete_manifest/2`
assumed that manifests were updated as their blocks were deleted.
This was based on an outdated view of GC. Now, the manifests
are copied into the GC bucket the GC daemon views and manipulates
them from this bucket.

This change makes it so that the manifests that are moved
to the GC bucket are:

1. the UUIDs passed into the GC function to begin with
2. any `pending_delete` manifests that appear to have been abandoned.
This means manifests that were put into the `pending_delete` state
but were never (or likely never) successfully moved to the GC bucket
and marked as `scheduled_delete`.