Clean up unused parts of multipart uploads
If a part isn't specified in a complete multipart upload
operation, then it must be deleted.  In order to accomodate
this, we use ?MULTIPART_MANIFEST.cleanup_parts to store
the list of parts to delete.  Once those parts have been
successfully added to the GC queue/thingie, then the
property `multipart_clean` is added to the props list.

If we do a get or delete of a multipart object and we do
not see the `multipart_clean` prop, then we attempt to
re-add the cleanup_parts list to the GC queue/thingie.

Undo the experiment of changing the nature of the
`multipart` property when the manifest transitions from
writing -> active: now we always keep it a ?MULTIPART_MANIFEST
record.