Add max size to list objects ETS cache
This commit implements an approximate max size constraint
to the list objects ETS cache. The constraint is approximate because
there is a small race condition between a process asking for the size
of the table and then adding new data. Since the max table size
is likely to be at least an order of magnitude bigger than an individual
element, the short race is not seen as a big deal. The race _could_
be alleviated by having all writes go through (or be given permission)
through the `riak_cs_list_objects_ets_cache` `gen_server`.
The size that a particular keylist of going to add to the table
is also approximated by multiplying the number of keys by a constant.

This commit also adds a minimum number of keys requirement to the cache.
The theory here being that doing a fresh keylist each time for small
buckets is not prohibitive and eliminates stale key listings.