Expire list objects ETS cache on WM exit too
Previously, the list objects ETS cache would expire solely
based on time. The cache key included the pid of the webmachine
process, so when the process exited (usually from the HTTP connection
closing), the cache was no longer going to be used. The timer might
not clean up the key for a while. This commit adds monitoring of
the specified pid to be another way that the cache key can be cleaned
up. Furthermore, two two expiration mechanisms clean each other up.
Put another way, if the timer goes off, the monitor is cancelled. If
a 'DOWN' message is received, the timer is cancelled. This code
_should_ handle all cases where a timer/monitor has been cancelled,
but the message already made it the mailbox.