Fix error in calculation of RFC1123 dates
Fixes #402

Fix a miscalculation when using RFC1123 dates. The problem stemmed
from the fact the httpd_util:rfc1123_date/1 expects input to be in
local time and attempts to convert it to GMT. All timestamps are
created in GMT so this is a change to convert the GMT time to local
time before calling httpd_util:rfc1123/1.