Use a separate webmachine instance for admin API
Feature: #265

Allow the admin API to use a separate IP and port number to more
easily restrict access to the administrative functionality. The user
may define admin_ip and admin_port distinct from cs_ip and cs_port in
the app.config file. If not specified the default behavior is for
admin and object storage resources to use the same IP address and port
number.