Add limited support for PUT copy + Content-Length: 0, to support S3FS
The S3FS FUSE client uses the feature of PUT copy + "Content-Length: 0"
to update only the metadata of an S3 object.  This metadata update is
used to support inode-modifying operations such as "touch", "chmod",
and "chown".

Add a new member to the #key_context record:

    update_metadata=false :: boolean()

In valid_entity_length(), check to see if "Content-Length: 0" and also
the "x-amz-copy-source" header is present and (if yes) if its value
is exactly equal to the object's path.

In accept_body(), check the new #key_context.update_metadata flag.
If true, update the ACL creation time (so that
riak_cs_manifest_resolution.erl logic can pick the newest version of the
manifest) and then write the updated manifest using
riak_cs_utils:set_object_acl().