Use named WM router table instead of starting them
* Update to work with basho/webmachine#128
* Builds on #265

@seancribbs' Webmachine change uses a single Webmachine router
(the way it used to be) and uses multiple keys in an ETS table
to support multiple dispatch routes. This change removes the custom
routers from the supervisor and adds a `router_name` to the two
webmachine options. Only one of them is used if the admin and web
config share the same interface.