Fix issue with request query string canonicalization
Fixes #441

The canonicalize_qs/2 function in the riak_cs_s3_auth module calls
canonicalize_qs/1 instead of canonicalize_qs/2 in the case where the
query parameter is not a valid S3 subresource. This leads to cases
where some query information can be lost and cause authentication
failure. This changes that code to correctly use canonicalize_qs/2 for
those cases.