Add test for stats resource and refactor cluster setup
Fixes #453

Add riak_test module for the stats resource. Also refactor the cluster
setup in the rtcs module so that the admin user credentials are set
and the riak_cs and stanchion nodes are restarted for those settings
to take effect. The rtcs:setup function return value has been changed
to also return the admin account info and allows for some duplicated
code to be eliminated in most of the test modules. These changes
resolve a race condition where user creation is attempted prior to the
riak_cs and stanchion applications being fully started.