Fix bucket ACL update incompatabilities
Fixes #467

Fix two incompatibilities with bucket ACLs. Bucket owners should
always be granted all ACL permissions except READ even when not
explicitly granted. The first fix addresses the issue that bucket
owners were being denied any access that was not explicitily granted.
The second change fixes an issue with ACL updates when specified
display_name for Owner element and a Grant element for the bucket
owner do not match. This has the side-effect of avoiding a 2I query to
riak in the case where the bucket owner is also the grantee.