Ensure bucket names only contain valid characters
Fixes #357

Ensure that bucket names follow the limitations so that they are DNS
compliant. The complete list of restrictions is at:
http://docs.aws.amazon.com/AmazonS3/latest/dev/BucketRestrictions.html. Update
all riak_test modules to use valid bucket names for testing. Also
define a function clause to handle 405 error responses in the
webmachine error handler to better conform to the S3 response format.