Improve handling for case when proxy_get option is enabled in OSS build
Fixes #501

Add exception handling around the cluster id retrieval code in
riak_cs_utils so that if proxy_get is enabled in an OSS build it does
not cause an ugly exception to be thrown. Also proactively set
proxy_get to disabled in this case to avoid having to repeatedly
handle the same exception.