Hibernate access archiver process to force memory to be released
Fixes #532

The transfer of an ets table of access information to the access
archiver process can result in excessive memory usage. The archiver
process experiences heap growth when archiving the access data and the
erlang vm seems unable to automatically free the memory
afterwards. Mitigate this by putting the archiver process into a state
of hibernation when it transitions to the idle state. This forces a gc
for the process before and after hibernation. This is acceptable
because the archiver process spends the majority of its time in the
idle state.