Add an abstraction between old and new list keys
The API is the same once you've called `start_link`. So this code has
you pass in the greater or the args between them (the old one), and then
returns a pid using either the old or the new one, dependent on the
value of `fold_objects_for_list_keys`.