Address memory consumption issue with access data archiving
Fixes #552

The access archiver is a long running process responsible for
archiving access information. When it receives access log data from a
log process that accumulates the data it can accumulate memory that is
not reclaimed by gc in a timely manner. Address this problem changing
the role of the long running process to be a managerial and have the
actual archiving done by temporary processes. This avoids the problem
of the long-running process consuming excessive amounts of memory and
ensure that memory used by the processes accumulating the access log
information is freed in a timely manner.