Properly include riak-cs-ee dependencies in packages
The `rebar.config.script` defines extra dependencies to be
included when building the enterprise version of riak-cs.
These dependencies were not being included in the 1.3 packages.

This fix simply exports the variable used `RIAK_CS_EE_DEPS`
to the packaging steps so subsequent calls to rebar still
include all the enterprise settings.

Fixes: #557 and basho/node_package#48