Integrate the ceph s3-tests suite into automated client tests
Add the ceph s3-tests suite to the set of automated python client
tests. This includes a script to create 2 user accounts and populate a
config file with information for those users. Currently the master
branch of the basho fork of the test suites is used, but in the future
this should be changed to use a tag. Also for now only the
functional test suite is run.