Fix the EE build check so that OSS builds do not generate errors
Fixes #568

The call to os:getenv to check for an EE build in rebar.config.script
sometimes returns an empty list instead of false. Change the check to
treat both of these cases as indicative of an OSS build.  This avoids
OSS builds attempting to check out private repos that they cannot
access and generating errors.