Stop all application nodes after riak_test run
Updates so that the riak_test modules use the rt_cs_dev test harness
instead of the rtdev Riak harness. This has changes to properly stop
all of the CS and Stanchion nodes. It also requires some configuration
changes. See the updates to the README in the riak_test subdirectory
for more details.