Move resolv_robj_siblings/1 to riak_kv_fs2_backend, to fix handoff problem
I noticed that handoff kept failing due to an error on the receiving
vnode (which caused "... failed because of error:{badmatch,{error,closed}} ..."
messages on the sending side).  The problem was caused by siblings in
the fs2's put_object() func, where one of the siblings was a tombstone.

resolv_robj_siblings/1 is the function that can figure that kind of
situation out correctly, so I moved it to the fs2 backend to try to
minimize external Riak-CS-related module dependencies.