Fix query parameter handling for pre-signed multipart upload URLs
Fixes #587

Pre-signed requests pertaining to multipart uploads fail on user
account look-up and authentication signature calculation because the
query parameters such as Signature and AWSAccessKeyId are improperly
removed during URL rewriting. Change the rewriting rules so that these
query parameters are not removed and address holes in the unit test
coverage of the S3 URL rewrite module.