Use S3 format for final etag of multipart uploads
Fixes #585

S3 appends a suffix to the etag header value for downloads of objects
that are uploaded using a multipart upload. This suffix is composed of
a dash and an integer representing the total number of parts that
comprised the upload. The etag value of an object uploaded using
multipart upload is not an MD5 checksum of the entire object and some
client libraries use the presence of the multipart upload etag suffix
as a means to determine whether to use the etag to validate the
integrity of the object download or if the failure of this integrity
check should be a cause for failure. Change RiakCS to also use this
etag suffix for multipart uploads to avoid spurious error messages or
failures from any client libraries that expect this suffix.