Never rewind when skipping keys in keylist
Fixes
[this](https://github.com/basho/riak_cs/pull/655#issuecomment-23309088)
issue comment. When skipping past a common prefix, we used to mistakenly
skip past only the last common prefix in the state record. There could,
however, be a later-sorted common prefix in the result set. So we now
skip past the exact common prefix that the last result is (if it is
one), instead of the last common prefix in the common prefixes field.