Handle undefined `props` in manifests
Prior to Riak CS 1.3.0 (667319391c4ef3d5c0d757e11952a5202cc64050 in
particularly), manifests were written with `undefined` in their `props`
field, despite the record definition setting it to `[]`. This change
fixes the manifests as they are read, and 'upgraded'. That way, the rest
of the code-base can assume the field is a proper proplist.