Address performance issues when exporting XML documents
Fixes #669

Address the repeated calls to error_handler:undefined_function that
results in a call to the code_server and is a performance bottleneck.
This commit adds a dependency on a fork of the xmerl library extracted
from the otp bundle. This fork has been modified so that a default
callback function can be specified for handling an element instead of
the only option being to attempt to call a function matching the
element name first and then calling the default element handler.

This commit also changes the riak_cs_xml module to pass in the tuples
of {'#element#', ElementName} instead of just ElementName to directly
use the default element callback function. This avoids having to
implement a callback function for each and every element name and
avoid the penalty of hitting the code_server numerous times when
generating a response for a user request.