Make logging in bucket listing map function lager version agnostic
The parse transform for lager changed between lager 1.2.X and lager
2.0.0. This causes problems listing the contents of buckets when using
Riak CS 1.4 or later with older versions of Riak because the
map_keys_and_manifests function is run on Riak nodes as part of a
MapReduce job. That function contains a lager call that is rewritten
to use a lager module that did not exist in versions prior to 2.0.0
and when run on Riak nodes that still uses an earlier lager version it
can result in errors. The errors happen when the bucket that a list
operation is run against contains a deleted object. This commit
changes the lager call in the map_keys_and_manifests function to use a
call that is not transformed and is safe for older Riak versions as
well as the 1.4.X releases.