Fix build of riak-cs-ee so RIAK_CS_EE_DEPS envvar gets passed properly
During build of debian packages, the environment is cleaned by debuild
by setting all environmental variables to "".  This fact made the
passing of RIAK_CS_EE_DEPS variable work on accident until the `[]->false`
pattern was added to the rebar.config.script.  Though adding that
pattern fixed the OSS build, it did break debian package builds.

This commit fixes this by removing that check, but also not
automatically exporting the RIAK_CS_EE_DEPS makefile since an export
will export an empty variable and the rebar.config.script can now
trigger on an empty setting again.

Calling `unset RIAK_CS_EE_DEPS` is the only way to build an OSS
build after previously exporting the variable for riak-cs-ee.

In addtion a flag is passed into the debian's debuild so when
it RIAK_CS_EE_DEPS is available even when the environment is cleaned.