Add option to allow a delay for the initial GC collection
Add an option that allows the initial run of the GC daemon to be delay
by some number of seconds. This provides a convenient way for the GC
daemon to be run on multiple nodes without overlapping the collection
and cleanup effort among nodes. The option is called initial_gc_delay
and is specified in seconds. The value only applies to the initial
collection run when the riak_cs node is started.