Make GC key suffix max value configurable by application env
The default value is 256. In terms of scalability, larger value is better.
On the other hand, when there are more keys in GC bucket, listing keys takes
longer. The default value 256 is conservative, safe side.

If GC mechanism will be improved (streaming/multi-threading), we may safely
adopt larger value, for example 1024 or more.