Change GC daemon to optionally use paginated 2I queries
Change the GC daemone to optionally use paginated 2I queries to avoid
system stress and timeouts when the amount of garbage to be collected
is large. The `paginated_indexes` boolean option is used to enable the
use of paginated queries and may be used when the version of Riak
being used is >= 1.4.0. The default for the option is false. There is
also an option, `gc_batch_size`, to configure the amount of results
that should be requested by the 2I query at one time. The default for
this option is 1000.